:tractor: relocate mkdir_p function to external dependencies folder
