+++ moban/core/utils.py
@@ -1,21 +1,9 @@
-import os
-import errno
 import logging
 
 from moban import constants, exceptions
 from moban.externals import file_system
 
 LOG = logging.getLogger(__name__)
-
-
-def mkdir_p(path):
-    try:
-        os.makedirs(path)
-    except OSError as exc:  # Python >2.5
-        if exc.errno == errno.EEXIST and os.path.isdir(path):
-            pass
-        else:
-            raise
 
 
 def verify_the_existence_of_directories(dirs):
