+++ moban/externals/buffered_writer.py
@@ -3,7 +3,6 @@ import sys
 
 import fs
 import fs.path
-from moban.core import utils
 from moban.externals import file_system
 
 PY2 = sys.version_info[0] == 2
@@ -44,6 +43,6 @@ def write_file_out(filename, content):
     if not file_system.is_zip_alike_url(filename):
         dest_folder = os.path.dirname(filename)
         if dest_folder:
-            utils.mkdir_p(dest_folder)
+            file_system.mkdir_p(dest_folder)
 
     file_system.write_bytes(filename, content)
